+++ sqlite3_web/lib/sqlite3_web.dart
@@ -1,2 +1,3 @@
 export 'src/types.dart';
 export 'src/database.dart' hide lockTokenFromId, lockTokenToId;
+export 'src/worker_connector.dart';